распределил модели по папкам добавил поле image в модели part
